requirements.txt
streamlit
numpy
pillow
imagehash